# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Gestão de Usuários",
        "description": "Esta API de Gestão de Usuários foi construida com o intuito de servir como exemplo de aplicação de uma Arquitetura Limpa (Clean Architecture)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": ""
        }
    ],
    "paths": {
        "/usuarios": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Busca um ou mais usuários de acordo com os filtros informados.",
                "operationId": "97264a8233714b17bbd6a7a744d8fa00",
                "responses": {
                    "200": {
                        "description": "Usuários encontrados com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID do usuário",
                                                "type": "integer"
                                            },
                                            "nome": {
                                                "description": "Nome do usuário",
                                                "type": "string"
                                            },
                                            "idade": {
                                                "description": "Idade do usuário",
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "description": "Data de criação",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "description": "Data de última atualização",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "description": "Campo",
                                                "type": "string"
                                            },
                                            "message": {
                                                "description": "Mensagem",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Servidor não pode atender à solicitação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        },
                                        "timestamp": {
                                            "description": "Data da ocorrência do erro",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Cria um novo usuário.",
                "operationId": "6906c2c7a27c047636548e5ff835b01a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "idade"
                                ],
                                "properties": {
                                    "nome": {
                                        "description": "Nome do usuário",
                                        "type": "string"
                                    },
                                    "idade": {
                                        "description": "Idade do usuário",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso.",
                        "headers": {
                            "Location": {
                                "description": "URL do novo usuário",
                                "schema": {
                                    "type": "string",
                                    "example": "/usuarios/{id}"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do usuário",
                                            "type": "integer"
                                        },
                                        "nome": {
                                            "description": "Nome do usuário",
                                            "type": "string"
                                        },
                                        "idade": {
                                            "description": "Idade do usuário",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Data de criação",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Data de última atualização",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Campos inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "description": "Campo",
                                                "type": "string"
                                            },
                                            "message": {
                                                "description": "Mensagem",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Servidor não pode atender à solicitação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensagem de erro",
                                            "type": "string"
                                        },
                                        "timestamp": {
                                            "description": "Data da ocorrência do erro",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
...
}